Fix bug handling images with links.
Before this commit links with images would get corrupted: the image would
simply be removed.
This commit changes the link handling logic to not care about what's inside
the text in a link and simply copy it verbatim from the source markdown,
avoiding this and other similar problems were things are nested inside a link's
text.